+++ e2e/helpers/pages/admin/tags/TagsPage.ts
@@ -1,28 +1,5 @@
-import {Locator, Page, Request} from '@playwright/test';
+import {Locator, Page} from '@playwright/test';
 import {AdminPage} from '../AdminPage';
-
-// XXX: Remove these types and the mockTagsResponse method once we have proper
-// test isolation and no longer need to mock responses.
-interface Tag {
-    id: string;
-    name: string;
-    slug: string;
-    url: string;
-    description: string;
-}
-
-interface PaginatedResponse {
-    meta: {
-        pagination: {
-            page: number;
-            limit: number;
-            pages: number;
-            total: number;
-            next?: number;
-        };
-    };
-    tags: Tag[];
-}
 
 export class TagsPage extends AdminPage {
     readonly pageContent: Locator;
@@ -63,27 +40,5 @@ export class TagsPage extends AdminPage {
 
     getRowByTitle(title: string) {
         return this.tagListRow.filter({has: this.page.getByRole('link', {name: title, exact: true})});
-    }
-
-    // XXX: Remove once we have proper test isolation and don't need mocking
-    async mockTagsResponse(handler: (request: Request) => Promise<Partial<PaginatedResponse>>) {
-        await this.page.route('/ghost/api/admin/tags/*', async (route, request) => {
-            const tags = await handler(request);
-            await route.fulfill({
-                body: JSON.stringify({
-                    meta: {
-                        ...tags.meta,
-                        pagination: {
-                            page: 1,
-                            limit: 100,
-                            pages: 1,
-                            total: tags.meta?.pagination?.total ?? tags.tags?.length ?? 0,
-                            ...tags.meta?.pagination
-                        }
-                    },
-                    tags: tags.tags ?? []
-                })
-            });
-        });
     }
 }
